+++ src/RoadRegistry.BackOffice/Messages/AddRoadSegmentToNationalRoad.cs
@@ -1,6 +1,8 @@
-﻿namespace RoadRegistry.BackOffice.Messages;
+namespace RoadRegistry.BackOffice.Messages;
 
-public class AddRoadSegmentToNationalRoad
+using Be.Vlaanderen.Basisregisters.EventHandling;
+
+public class AddRoadSegmentToNationalRoad : IMessage
 {
     public string Number { get; set; }
     public int SegmentId { get; set; }