+++ src/RoadRegistry.BackOffice/Messages/RoadNetworkExtractGeometry.cs
@@ -4,5 +4,6 @@ public class RoadNetworkExtractGeometry
 {
     public Polygon[] MultiPolygon { get; set; }
     public Polygon Polygon { get; set; }
+    public string WKT { get; set; }
     public int SpatialReferenceSystemIdentifier { get; set; }
 }