+++ src/RoadRegistry.Hosts/WellknownConnectionNames.cs
@@ -18,8 +18,10 @@ public static class WellknownConnectionNames
     public const string SnapshotsAdmin = "SnapshotsAdmin";
     public const string SyndicationProjections = "SyndicationProjections";
     public const string SyndicationProjectionsAdmin = "SyndicationProjectionsAdmin";
-    public const string WfsProjections = "WfsProjections";
-    public const string WfsProjectionsAdmin = "WfsProjectionsAdmin";
     public const string WmsProjections = "WmsProjections";
     public const string WmsProjectionsAdmin = "WmsProjectionsAdmin";
+    public const string WfsProjections = "WfsProjections";
+    public const string WfsProjectionsAdmin = "WfsProjectionsAdmin";
+    public const string StreetNameConsumer = "StreetNameConsumer";
+    public const string StreetNameConsumerAdmin = "StreetNameConsumerAdmin";
 }