+++ test/RoadRegistry.BackOffice.MessagingHost.Sqs.Tests/Startup.cs
@@ -47,7 +47,7 @@ public class Startup : TestStartup
             }));
     }
 
-    public override void ConfigureContainer(ContainerBuilder builder)
+    protected override void ConfigureContainer(ContainerBuilder builder)
     {
         builder.RegisterModule<MediatorModule>();
 
@@ -68,7 +68,7 @@ public class Startup : TestStartup
         //}).As<ProductContext>();
     }
 
-    public override void ConfigureServices(HostBuilderContext hostBuilderContext, IServiceCollection services)
+    protected override void ConfigureServices(HostBuilderContext hostBuilderContext, IServiceCollection services)
     {
         services
             .AddSingleton(_ => new SqsOptions("", "", RegionEndpoint.EUWest1))